add multi-stage docker build
there were 3 index.html files,
in the root, in the templates and in the src folder;
only the one in the src folder is important since
it will generate the one in templates, the one in the root folder is
irrelevant.

Also fix the .gitignore for the above and for the node_moudules
and python cache.